Feat(e2e): Create a utility function to auto-fill forms.
